Add Red Hat Partner certification for v10 (Newton)
addresses issue 247 which appears to be a duplicate of 237

Jodie, make sure to check this as I think I have the formatting of the table correct, but I can not view it in the preview pane so not positive. and the markdown viewer I passed it into didn't like the table.